Formated code using forge fmt

+++ script/HelperConfig.sol
@@ -6,8 +6,8 @@ import {Script, console2} from "forge-std/Script.sol";
 import {VRFCoordinatorV2_5Mock} from "@chainlink/contracts/src/v0.8/vrf/mocks/VRFCoordinatorV2_5Mock.sol";
 
 abstract contract CodeConstants {
-    uint public constant LOCAL_CHAIN_ID = 31337;
-    uint public constant ETH_SEPOLIA_CHAIN_ID = 11155111;
+    uint256 public constant LOCAL_CHAIN_ID = 31337;
+    uint256 public constant ETH_SEPOLIA_CHAIN_ID = 11155111;
 
     uint96 public constant MOCK_BASE_FEE = 0.00025 ether;
     uint96 public constant MOCK_GAS_PRICE = 1e9;
@@ -20,8 +20,8 @@ contract HelperConfig is Script, CodeConstants {
 
     /* Type declarations */
     struct NetworkConfig {
-        uint entranceFee;
-        uint interval;
+        uint256 entranceFee;
+        uint256 interval;
         address vrfCoordinator;
         bytes32 keyHash;
         uint256 subscriptionId;
@@ -38,9 +38,7 @@ contract HelperConfig is Script, CodeConstants {
         return getConfigByChainId(block.chainid);
     }
 
-    function getConfigByChainId(
-        uint chainId
-    ) private returns (NetworkConfig memory) {
+    function getConfigByChainId(uint256 chainId) private returns (NetworkConfig memory) {
         if (chainId == LOCAL_CHAIN_ID) {
             return getOrCreateEthAnvilConfig();
         } else if (chainId == ETH_SEPOLIA_CHAIN_ID) {
@@ -50,10 +48,7 @@ contract HelperConfig is Script, CodeConstants {
         }
     }
 
-    function getOrCreateEthAnvilConfig()
-        private
-        returns (NetworkConfig memory)
-    {
+    function getOrCreateEthAnvilConfig() private returns (NetworkConfig memory) {
         if (localNetworkConfig.vrfCoordinator != address(0)) {
             return localNetworkConfig;
         }
@@ -61,11 +56,8 @@ contract HelperConfig is Script, CodeConstants {
         console2.log("Make sure this was intentional");
 
         vm.startBroadcast();
-        VRFCoordinatorV2_5Mock mockVrfCoordinator = new VRFCoordinatorV2_5Mock(
-            MOCK_BASE_FEE,
-            MOCK_GAS_PRICE,
-            MOCK_WEI_PER_UNIT_LINK
-        );
+        VRFCoordinatorV2_5Mock mockVrfCoordinator =
+            new VRFCoordinatorV2_5Mock(MOCK_BASE_FEE, MOCK_GAS_PRICE, MOCK_WEI_PER_UNIT_LINK);
         LinkToken linkToken = new LinkToken();
         vm.stopBroadcast();
 
@@ -84,16 +76,15 @@ contract HelperConfig is Script, CodeConstants {
     }
 
     function getEthSepliaConfig() private pure returns (NetworkConfig memory) {
-        return
-            NetworkConfig({
-                entranceFee: 0.0025 ether,
-                interval: 60,
-                vrfCoordinator: 0x9DdfaCa8183c41ad55329BdeeD9F6A8d53168B1B,
-                keyHash: 0x787d74caea10b2b357790d5b5247c2f63d1d91572a9846f780606e4d953677ae,
-                callbackGasLimit: 500000,
-                subscriptionId: 0,
-                link: 0x779877A7B0D9E8603169DdbD7836e478b4624789,
-                account: 0xA2358D11ABf324976095c7E6cfF07D6B5DE9fA09
-            });
+        return NetworkConfig({
+            entranceFee: 0.0025 ether,
+            interval: 60,
+            vrfCoordinator: 0x9DdfaCa8183c41ad55329BdeeD9F6A8d53168B1B,
+            keyHash: 0x787d74caea10b2b357790d5b5247c2f63d1d91572a9846f780606e4d953677ae,
+            callbackGasLimit: 500000,
+            subscriptionId: 0,
+            link: 0x779877A7B0D9E8603169DdbD7836e478b4624789,
+            account: 0xA2358D11ABf324976095c7E6cfF07D6B5DE9fA09
+        });
     }
 }
