+++ script/Interactions.s.sol
@@ -9,74 +9,58 @@ import {DevOpsTools} from "lib/foundry-devops/src/DevOpsTools.sol";
 import {VRFCoordinatorV2_5Mock} from "@chainlink/contracts/src/v0.8/vrf/mocks/VRFCoordinatorV2_5Mock.sol";
 
 contract CreateSubscription is Script {
-    function createSubscriptionUsingConfig() private returns (uint, address) {
+    function createSubscriptionUsingConfig() private returns (uint256, address) {
         HelperConfig helperConfig = new HelperConfig();
         address vrfCoordinator = helperConfig.getConfig().vrfCoordinator;
         address account = helperConfig.getConfig().account;
 
-        (uint subId, ) = createSubscription(vrfCoordinator, account);
+        (uint256 subId,) = createSubscription(vrfCoordinator, account);
         return (subId, vrfCoordinator);
     }
 
-    function createSubscription(
-        address vrfCoordinator,
-        address account
-    ) public returns (uint, address) {
+    function createSubscription(address vrfCoordinator, address account) public returns (uint256, address) {
         console2.log("Creating subscription on chain Id: ", block.chainid);
 
         vm.startBroadcast(account);
-        uint subscriptionId = VRFCoordinatorV2_5Mock(vrfCoordinator)
-            .createSubscription();
+        uint256 subscriptionId = VRFCoordinatorV2_5Mock(vrfCoordinator).createSubscription();
         vm.stopBroadcast();
 
         console2.log("Your subscription Id is: ", subscriptionId);
-        console2.log(
-            "Please update you subscription Id in your HelperConfig.s.sol"
-        );
+        console2.log("Please update you subscription Id in your HelperConfig.s.sol");
         return (subscriptionId, vrfCoordinator);
     }
 
-    function run() external returns (uint, address) {
+    function run() external returns (uint256, address) {
         return createSubscriptionUsingConfig();
     }
 }
 
 contract FundSubscription is Script, CodeConstants {
-    uint public constant FUND_AMOUNT = 10 ether;
+    uint256 public constant FUND_AMOUNT = 10 ether;
 
     function fundSubscriptionUsingConfig() private {
         HelperConfig helperConfig = new HelperConfig();
         address vrfCoordinator = helperConfig.getConfig().vrfCoordinator;
-        uint subId = helperConfig.getConfig().subscriptionId;
+        uint256 subId = helperConfig.getConfig().subscriptionId;
         address linkToken = helperConfig.getConfig().link;
         address account = helperConfig.getConfig().account;
         fundSubscription(vrfCoordinator, subId, linkToken, account);
     }
 
-    function fundSubscription(
-        address vrfCoordinator,
-        uint subscriptionId,
-        address linkToken,
-        address account
-    ) public {
+    function fundSubscription(address vrfCoordinator, uint256 subscriptionId, address linkToken, address account)
+        public
+    {
         console2.log("Funding subscription: ", subscriptionId);
         console2.log("Using vrfCoordinator: ", vrfCoordinator);
         console2.log("On Chain Id: ", block.chainid);
 
         if (block.chainid == LOCAL_CHAIN_ID) {
             vm.startBroadcast();
-            VRFCoordinatorV2_5Mock(vrfCoordinator).fundSubscription(
-                subscriptionId,
-                FUND_AMOUNT
-            );
+            VRFCoordinatorV2_5Mock(vrfCoordinator).fundSubscription(subscriptionId, FUND_AMOUNT);
             vm.stopBroadcast();
         } else {
             vm.startBroadcast(account);
-            LinkToken(linkToken).transferAndCall(
-                vrfCoordinator,
-                FUND_AMOUNT,
-                abi.encodePacked(subscriptionId)
-            );
+            LinkToken(linkToken).transferAndCall(vrfCoordinator, FUND_AMOUNT, abi.encodePacked(subscriptionId));
             vm.stopBroadcast();
         }
     }
@@ -90,40 +74,26 @@ contract AddConsumer is Script {
     function addConsumerUsingConfig(address mostRecentlyDeployed) private {
         HelperConfig helperConfig = new HelperConfig();
         address vrfCoordinator = helperConfig.getConfig().vrfCoordinator;
-        uint subscriptionId = helperConfig.getConfig().subscriptionId;
+        uint256 subscriptionId = helperConfig.getConfig().subscriptionId;
         address account = helperConfig.getConfig().account;
 
-        addConsumer(
-            mostRecentlyDeployed,
-            vrfCoordinator,
-            subscriptionId,
-            account
-        );
+        addConsumer(mostRecentlyDeployed, vrfCoordinator, subscriptionId, account);
     }
 
-    function addConsumer(
-        address contractToAddToVrf,
-        address vrfCoordinator,
-        uint subscriptionId,
-        address account
-    ) public {
+    function addConsumer(address contractToAddToVrf, address vrfCoordinator, uint256 subscriptionId, address account)
+        public
+    {
         console2.log("Adding consumer contract: ", contractToAddToVrf);
         console2.log("To vrfCoordinator: ", vrfCoordinator);
         console2.log("On Chain Id: ", block.chainid);
 
         vm.startBroadcast(account);
-        VRFCoordinatorV2_5Mock(vrfCoordinator).addConsumer(
-            subscriptionId,
-            contractToAddToVrf
-        );
+        VRFCoordinatorV2_5Mock(vrfCoordinator).addConsumer(subscriptionId, contractToAddToVrf);
         vm.stopBroadcast();
     }
 
     function run() external {
-        address mostRecentlyDeployed = DevOpsTools.get_most_recent_deployment(
-            "Raffle",
-            block.chainid
-        );
+        address mostRecentlyDeployed = DevOpsTools.get_most_recent_deployment("Raffle", block.chainid);
         addConsumerUsingConfig(mostRecentlyDeployed);
     }
 }
