+++ src/Raffle.sol
@@ -1,72 +1,62 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.28;
 
-// Layout of Contract:
-// version
-// imports
-// errors
-// interfaces, libraries, contracts
-// Type declarations
-// State variables
-// Events
-// Modifiers
-// Functions
-
-// Layout of Functions:
-// constructor
-// receive function (if exists)
-// fallback function (if exists)
-// external
-// public
-// internal
-// private
-// internal & private view & pure functions
-// external & public view & pure functions
-
 import {VRFV2PlusClient} from "@chainlink/contracts/src/v0.8/vrf/dev/libraries/VRFV2PlusClient.sol";
 import {VRFConsumerBaseV2Plus} from "@chainlink/contracts/src/v0.8/vrf/dev/VRFConsumerBaseV2Plus.sol";
 
 contract Raffle is VRFConsumerBaseV2Plus {
-    /* Errors */
+
+    //////////////
+    /// Errors ///
+    //////////////
+
     error Raffle__RaffleNotOpen();
     error Raffle__TransferFailed();
     error Raffle__SendMoreToEnterRaffle();
-    error Raffle__UpkeepNotNeeded(
-        uint balance,
-        uint playersLength,
-        uint raffleState
-    );
+    error Raffle__UpkeepNotNeeded(uint256 balance, uint256 playersLength, uint256 raffleState);
 
-    /* Type declarations */
+    /////////////////////////
+    /// Type declarations ///
+    /////////////////////////
+
     enum RaffleState {
         OPEN,
         CALCULATING
     }
 
-    /* State variables */
+    ///////////////////////
+    /// State variables ///
+    ///////////////////////
+
     uint32 private constant NUM_WORDS = 1;
     uint16 private constant REQUEST_CONFIRMATIONS = 3;
 
-    uint private immutable i_interval;
+    uint256 private immutable i_interval;
     bytes32 private immutable i_keyHash;
-    uint private immutable i_entranceFee;
-    uint private immutable i_subscriptionId;
+    uint256 private immutable i_entranceFee;
+    uint256 private immutable i_subscriptionId;
     uint32 private immutable i_callbackGasLimit;
 
-    uint private s_lastTimeStamp;
+    uint256 private s_lastTimeStamp;
     address private s_recentWinner;
     RaffleState private s_raffleState;
     address payable[] private s_players;
 
-    /* Events */
+    //////////////
+    /// Events ///
+    //////////////
+
     event WinnerPicked(address indexed winner);
     event RaffleEntered(address indexed player);
-    event RequestedRaffleWinner(uint indexed requestId);
+    event RequestedRaffleWinner(uint256 indexed requestId);
 
-    /* Constructor */
+    ///////////////////
+    /// Constructor ///
+    ///////////////////
+
     constructor(
-        uint entranceFee,
-        uint interval,
+        uint256 entranceFee,
+        uint256 interval,
         address vrfCoordinator,
         bytes32 keyHash,
         uint256 subscriptionId,
@@ -82,7 +72,10 @@ contract Raffle is VRFConsumerBaseV2Plus {
         s_lastTimeStamp = block.timestamp;
     }
 
-    /* Functions */
+    //////////////////////////
+    /// External Functions ///
+    //////////////////////////
+
     function enterRaffle() external payable {
         require(msg.value >= i_entranceFee, Raffle__SendMoreToEnterRaffle());
         require(s_raffleState == RaffleState.OPEN, Raffle__RaffleNotOpen());
@@ -90,40 +83,31 @@ contract Raffle is VRFConsumerBaseV2Plus {
         emit RaffleEntered(msg.sender);
     }
 
-    function performUpkeep() external returns (uint) {
-        (bool upkeepNeeded, ) = checkUpkeep();
-        require(
-            upkeepNeeded,
-            Raffle__UpkeepNotNeeded(
-                address(this).balance,
-                s_players.length,
-                uint(s_raffleState)
-            )
-        );
+    function performUpkeep() external returns (uint256) {
+        (bool upkeepNeeded,) = checkUpkeep();
+        require(upkeepNeeded, Raffle__UpkeepNotNeeded(address(this).balance, s_players.length, uint256(s_raffleState)));
 
         s_raffleState = RaffleState.CALCULATING;
-        VRFV2PlusClient.RandomWordsRequest memory request = VRFV2PlusClient
-            .RandomWordsRequest({
-                keyHash: i_keyHash,
-                subId: i_subscriptionId,
-                requestConfirmations: REQUEST_CONFIRMATIONS,
-                callbackGasLimit: i_callbackGasLimit,
-                numWords: NUM_WORDS,
-                extraArgs: VRFV2PlusClient._argsToBytes(
-                    VRFV2PlusClient.ExtraArgsV1({nativePayment: false})
-                )
-            });
+        VRFV2PlusClient.RandomWordsRequest memory request = VRFV2PlusClient.RandomWordsRequest({
+            keyHash: i_keyHash,
+            subId: i_subscriptionId,
+            requestConfirmations: REQUEST_CONFIRMATIONS,
+            callbackGasLimit: i_callbackGasLimit,
+            numWords: NUM_WORDS,
+            extraArgs: VRFV2PlusClient._argsToBytes(VRFV2PlusClient.ExtraArgsV1({nativePayment: false}))
+        });
 
         uint256 requestId = s_vrfCoordinator.requestRandomWords(request);
         emit RequestedRaffleWinner(requestId);
         return requestId;
     }
 
-    function fulfillRandomWords(
-        uint256 /*requestId*/,
-        uint256[] calldata randomWords
-    ) internal override {
-        uint indexOfWinner = randomWords[0] % s_players.length;
+    //////////////////////////
+    /// Internal Functions ///
+    //////////////////////////
+
+    function fulfillRandomWords(uint256, /*requestId*/ uint256[] calldata randomWords) internal override {
+        uint256 indexOfWinner = randomWords[0] % s_players.length;
         address payable recentWinner = s_players[indexOfWinner];
         s_recentWinner = recentWinner;
 
@@ -132,9 +116,13 @@ contract Raffle is VRFConsumerBaseV2Plus {
         s_players = new address payable[](0);
         emit WinnerPicked(s_recentWinner);
 
-        (bool success, ) = recentWinner.call{value: address(this).balance}("");
+        (bool success,) = recentWinner.call{value: address(this).balance}("");
         require(success, Raffle__TransferFailed());
     }
+
+    //////////////////////////////////////////
+    /// External and Public View Functions ///
+    //////////////////////////////////////////
 
     function checkUpkeep() public view returns (bool, bytes memory) {
         bool timeHasPassed = (block.timestamp - s_lastTimeStamp) >= i_interval;
@@ -146,18 +134,15 @@ contract Raffle is VRFConsumerBaseV2Plus {
         return (upkeepNeeded, "0x");
     }
 
-    /* Get functions */
-    function getPlayer(
-        uint indexOfPlayer
-    ) external view returns (address payable) {
+    function getPlayer(uint256 indexOfPlayer) external view returns (address payable) {
         return s_players[indexOfPlayer];
     }
 
-    function getEntranceFee() external view returns (uint) {
+    function getEntranceFee() external view returns (uint256) {
         return i_entranceFee;
     }
 
-    function getInterval() external view returns (uint) {
+    function getInterval() external view returns (uint256) {
         return i_interval;
     }
 
@@ -165,7 +150,7 @@ contract Raffle is VRFConsumerBaseV2Plus {
         return i_keyHash;
     }
 
-    function getSubscriptionId() external view returns (uint) {
+    function getSubscriptionId() external view returns (uint256) {
         return i_subscriptionId;
     }
 
@@ -181,7 +166,7 @@ contract Raffle is VRFConsumerBaseV2Plus {
         return s_raffleState;
     }
 
-    function getLastTimeStamp() external view returns (uint) {
+    function getLastTimeStamp() external view returns (uint256) {
         return s_lastTimeStamp;
     }
